# apps/js-calc-backend/app/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a calculator api",
      "version": "1.0.0",
      "title": "Prime Calculator",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "dennis@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "basePath": "/",
    "tags": [
      {
        "name": "api",
        "description": "Everything about calculations",
        "externalDocs": {
          "description": "Find out more",
          "url": "http://swagger.io"
        }
      },
      {
        "name": "ping",
        "description": "Everything about health",
        "externalDocs": {
          "description": "Find out more",
          "url": "http://swagger.io"
        }
      }
    ],
    "schemes": [
      "http",
      "https"
    ],
    "paths": {
      "/ping": {
        "get": {
          "tags": [
            "ping"
          ],
          "summary": "Check if server is alive",
          "description": "",
          "operationId": "ping",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "responses": {
            "500": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/api/calculation": {
        "post": {
          "tags": [
            "api"
          ],
          "summary": "Calculates the prime factors",
          "description": "",
          "operationId": "calculation",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "number",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful calculation",
              "schema": {
                "$ref": "#/definitions/Result"
              }
            },
            "500": {
              "description": "Internal error"
            }
          }
        }
      }
    },
    "definitions": {
      "Result": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Calculation result"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "host": {
            "type": "string",
            "description": "Host name"
          },
          "remote": {
            "type": "string",
            "description": "Caller ip"
          }
        },
        "xml": {
          "name": "Result"
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }